fixes for making it compile with XCode 8.2 and MacOS 10.11

+++ browser/HexFormatter.m
@@ -29,7 +29,7 @@
 
 + (NSString*)format:(long)number;
 {
-    return [NSString stringWithFormat: @"0x%02X", number];
+    return [NSString stringWithFormat: @"0x%02X", (unsigned int)number];
 }
